api v1 - move Authorization to its own file
move nested class Authorization to lib
rename setting category OAuth to Auth
