+++ app/controllers/api/base_controller.rb
@@ -1,5 +1,3 @@
-require 'oauth/client/action_controller_request'
-
 module Api
   #TODO: inherit from application controller after cleanup
   class BaseController < ActionController::Base
@@ -53,7 +51,7 @@ module Api
     end
 
     def authorize
-      auth = Authorize.new self
+      auth = Api::Authorization.new self
 
       unless auth.authenticate
         render_error('unauthorized', :status => :unauthorized, :locals => { :user_login => auth.user_login })
@@ -66,58 +64,6 @@ module Api
       end
 
       return true
-    end
-
-    class Authorize
-      attr_reader :controller, :user_login
-
-      def initialize(controller)
-        @controller = controller
-      end
-
-      def authenticate
-        unless SETTINGS[:login]
-          # We assume we always have a user logged in,
-          # if authentication is disabled, the user is the build-in admin account.
-          User.current = User.find_by_login("admin")
-        else
-          authorization_method = if controller.request.authorization =~ /^OAuth/
-                                   :oauth
-                                 else
-                                   :http_basic
-                                 end
-          User.current         ||= send(authorization_method) || (return false)
-        end
-
-        return true
-      end
-
-      def authorize
-        User.current.allowed_to?(
-            :controller => controller.params[:controller].gsub(/::/, "_").underscore,
-            :action     => controller.params[:action])
-      end
-
-      def http_basic
-        controller.authenticate_with_http_basic do |u, p|
-          @user_login = u
-          User.try_to_login(u, p)
-        end
-      end
-
-      def oauth
-        unless Setting['oauth_active'] &&
-            (OAuth::RequestProxy.proxy(controller.request).oauth_consumer_key == Setting['consumer_key'])
-          return nil
-        end
-
-        if OAuth::Signature.verify(controller.request, :consumer_secret => Setting['consumer_secret'])
-          # TODO find user by header
-          return User.find_by_login("admin")
-        else
-          return nil
-        end
-      end
     end
 
     def deny_access(details = nil)
